Add comments

+++ src/main/java/com/xtu/DB/vo/AllContestVO.java
@@ -9,7 +9,16 @@ import java.util.List;
  */
 @Data
 public class AllContestVO {
+    /**
+     * AllContestEntityVO列表
+     */
     private List<AllContestEntityVO> entityList;
-    private int total;
+    /**
+     * 总数
+     */
+    private long total;
+    /**
+     * 页数
+     */
     private int start;
 }
